Reorg into separate async_cdk module
