+++ airbyte-cdk/python/airbyte_cdk/sources/__init__.py
@@ -5,7 +5,6 @@
 import dpath.options
 
 from .abstract_source import AbstractSource
-from .abstract_source_async import AsyncAbstractSource
 from .config import BaseConfig
 from .source import Source
 
@@ -20,4 +19,4 @@ from .source import Source
 # this will not be thread-safe.
 dpath.options.ALLOW_EMPTY_STRING_KEYS = True
 
-__all__ = ["AbstractSource", "AsyncAbstractSource", "BaseConfig", "Source"]
+__all__ = ["AbstractSource", "BaseConfig", "Source"]
