+++ airbyte-cdk/python/airbyte_cdk/sources/async_cdk/abstract_source_async.py
@@ -24,7 +24,7 @@ from airbyte_cdk.models import (
 from airbyte_cdk.models import Type as MessageType
 from airbyte_cdk.sources.connector_state_manager import ConnectorStateManager
 from airbyte_cdk.sources.abstract_source import AbstractSource
-from airbyte_cdk.sources.streams import AsyncStream
+from airbyte_cdk.sources.async_cdk.streams.core_async import AsyncStream
 from airbyte_cdk.sources.streams.core import StreamData
 from airbyte_cdk.sources.streams.http.http import HttpStream
 from airbyte_cdk.sources.utils.record_helper import stream_data_to_airbyte_message
