+++ airbyte-cdk/python/airbyte_cdk/sources/async_cdk/streams/http/availability_strategy_async.py
@@ -8,9 +8,10 @@ from typing import Dict, Optional, Tuple
 
 import requests
 from aiohttp import ClientResponseError
-from airbyte_cdk.sources.streams import AsyncStream
+from airbyte_cdk.sources.async_cdk.streams.core_async import AsyncStream
+from airbyte_cdk.sources.async_cdk.streams.utils.stream_helper_async import get_first_record_for_slice, get_first_stream_slice
 from airbyte_cdk.sources.streams.http.availability_strategy import HttpAvailabilityStrategy
-from airbyte_cdk.sources.streams.utils.stream_helper_async import get_first_record_for_slice, get_first_stream_slice
+
 
 if typing.TYPE_CHECKING:
     from airbyte_cdk.sources import Source
