+++ airbyte-cdk/python/airbyte_cdk/sources/async_cdk/streams/http/exceptions_async.py
