+++ airbyte-cdk/python/airbyte_cdk/sources/async_cdk/streams/utils/stream_helper_async.py
@@ -5,7 +5,7 @@
 from typing import Any, Mapping, Optional
 
 from airbyte_cdk.models import SyncMode
-from airbyte_cdk.sources.streams import AsyncStream
+from airbyte_cdk.sources.async_cdk.streams.core_async import AsyncStream
 from airbyte_cdk.sources.streams.core import StreamData
