+++ airbyte-cdk/python/airbyte_cdk/sources/streams/__init__.py
@@ -4,6 +4,5 @@
 
 # Initialize Streams Package
 from .core import IncrementalMixin, Stream
-from .core_async import AsyncStream
 
-__all__ = ["IncrementalMixin", "Stream", "AsyncStream"]
+__all__ = ["IncrementalMixin", "Stream"]
